cs146s-learning-platform/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Flask-Login==0.6.3
Flask-Session==0.5.0
Werkzeug==2.3.7
python-dotenv==1.0.0
openai==1.3.0
markdown==3.5.1
pygments==2.16.1
requests==2.31.0
psutil==5.9.6
gunicorn==21.2.0
gevent==23.9.1

# Testing dependencies
pytest==7.4.2
pytest-cov==4.1.0
pytest-playwright==0.4.3
playwright==1.40.0